pa_mapformat in progress

+++ src/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: jcuzin <jcuzin@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/29 07:36:29 by jcuzin            #+#    #+#             */
-/*   Updated: 2024/04/02 22:17:38 by jcuzin           ###   ########.fr       */
+/*   Updated: 2024/04/02 23:05:32 by jcuzin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,8 +26,8 @@
 typedef struct s_player
 {
 	char		compass;
-	long long	xpos;
-	long long	ypos;
+	double		xpos;
+	double		ypos;
 	int			b_interact;
 	int			b_move;
 }				t_player;
@@ -36,8 +36,8 @@ typedef struct s_map
 {
 	int			fd;
 	char		**map;
-	size_t		xlen;
-	size_t		ylen;
+	double		xlen;
+	double		ylen;
 }				t_map;
 
 typedef struct s_data
